+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## [5.3.0-beta.4](https://github.com/Lint-Free-Technology/uix/compare/v5.3.0-beta.3...v5.3.0-beta.4) (2026-03-09)
+
+### ⭐ New Features
+
+* Add UIX parent variables to uix_tree() and uix_path(). ([6e027b3](https://github.com/Lint-Free-Technology/uix/commit/6e027b3e125a1a7520b752fabd5bc7e8091601ce))
+
+## [5.3.0-beta.3](https://github.com/Lint-Free-Technology/uix/compare/v5.3.0-beta.2...v5.3.0-beta.3) (2026-03-08)
+
+### 🐞 Bug Fixes
+
+* Applying Uix to section strategy causes it to not show and console errors ([899163e](https://github.com/Lint-Free-Technology/uix/commit/899163ee2dd07b6b060d46f91e148a54013ac60b))
+
+## [5.3.0-beta.2](https://github.com/Lint-Free-Technology/uix/compare/v5.3.0-beta.1...v5.3.0-beta.2) (2026-03-08)
+
+### ⭐ New Features
+
+* DOM inspecton helpers uix_tree() and uix_path() ([#78](https://github.com/Lint-Free-Technology/uix/issues/78)) ([51cabd5](https://github.com/Lint-Free-Technology/uix/commit/51cabd5fb5665131b6dd272b6269a03579d02a84))
+
+## [5.3.0-beta.1](https://github.com/Lint-Free-Technology/uix/compare/v5.2.2...v5.3.0-beta.1) (2026-03-08)
+
+### ⭐ New Features
+
+* Macro support for UIX templates - See https://uix.lf.technology/using/templates/[#macros](https://github.com/Lint-Free-Technology/uix/issues/macros) ([#75](https://github.com/Lint-Free-Technology/uix/issues/75)) ([3f6dd64](https://github.com/Lint-Free-Technology/uix/commit/3f6dd64a504d2760503b315803c9f08f02c6e034))
+
 ## [5.2.2](https://github.com/Lint-Free-Technology/uix/compare/v5.2.1...v5.2.2) (2026-03-07)
 
 ### 🐞 Bug Fixes